Add 2D processing functions to backend
Problem:  2D spectra could not be analysed with available cost
          functions. The use of 2D spectra should not be very common,
          but I plan to use the ASAP-HSQC as an example.
Solution: Write a function to read 2D data. Most of it is copied from
          penguins...

Also, I've changed the way the bounds are specified: now they're done in
a string style, similar to penguins. This should be more understandable
than the previous one where there is obvious confusion over which is
lower and which is upper.