Correctly handle TopSpin 4 raw data with dtypa=2, also endianness
Fixes #80, tested on AVB400.

I also added in a check for endianness for good measure. The code in
get1d_fid() is now very similar to the code in
penguins._1D_RawDataMixin._read_raw_data().

Problem:  TopSpin 4 introduced FIDs where numbers were stored as
          double-precision floats. The existing codebase only worked on
          TopSpin 3 data where FIDs were stored as 32-bit integers.
Solution: Check for the DTYPA parameter and read in the FID file
          accordingly.